fix(pa01): avoid always-true LED_CHARGING_START comparison in ledChargingInfo
Introduced in #6392: pa01 defines LED_CHARGING_START as 0, so
`i >= LED_CHARGING_START` with the loop's unsigned uint8_t i is
always true, triggering -Wtype-limits. Use a plain int loop variable
so the comparisons are ordinary signed comparisons instead of testing
a constant.

+++ radio/src/targets/pa01/battery_driver.cpp
@@ -216,7 +216,7 @@ void ledChargingInfo(uint16_t chargeState) {
     ledIdx = LED_CHARGING_START;
   else
     ledIdx += LED_CHARGING_START > LED_CHARGING_END ? -1 : 1;
-  for (uint8_t i = 0; i < LED_STRIP_LENGTH; i++) {
+  for (int i = 0; i < LED_STRIP_LENGTH; i++) {
     uint8_t green = 0;
     if (CHARGE_FINISHED == chargeState) {
       if (LED_CHARGING_START > LED_CHARGING_END && i <= LED_CHARGING_START && i >= LED_CHARGING_END)
